+++ src/common/util.rs
@@ -219,6 +219,12 @@ impl From<Vec<u8>> for HttpMockBytes {
     }
 }
 
+impl From<Box<[u8]>> for HttpMockBytes {
+    fn from(value: Box<[u8]>) -> Self {
+        Self(Bytes::from(value))
+    }
+}
+
 impl From<()> for HttpMockBytes {
     fn from(_: ()) -> Self {
         Self(Bytes::new())