+++ tests/misc/mod.rs
@@ -1,3 +1,4 @@
+mod body_conversion_tests;
 mod extensions_test;
 #[cfg(feature = "remote")]
 mod large_body_test;